% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.nces.colleges-package.R
\docType{package}
\name{dc.nces.colleges-package}
\alias{dc.nces.colleges}
\alias{dc.nces.colleges-package}
\title{dc.nces.colleges: Social Data Commons: Education & Training: Access to 4-year and 2-year Colleges, Vocational and Trade Schools}
\description{
Allows user to easily get access scores to 4- and 2-year colleges and trade schools data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Hanna Charankevich \email{hc2cc@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
